+++ eras/alonzo/impl/src/Cardano/Ledger/Alonzo/Rules/Utxo.hs
@@ -49,7 +49,7 @@ import Cardano.Ledger.Alonzo.Rules.Utxos (
   UtxosEnv (..),
  )
 import Cardano.Ledger.Alonzo.Scripts (ExUnits (..), pointWiseExUnits)
-import Cardano.Ledger.Alonzo.Tx (AlonzoEraTx (..), IsValid (..), totExUnits)
+import Cardano.Ledger.Alonzo.Tx (AlonzoEraTx (..), IsPhase2Valid (..), totExUnits)
 import Cardano.Ledger.Alonzo.TxBody (
   AllegraEraTxBody (..),
   AlonzoEraTxBody (..),
@@ -573,8 +573,8 @@ utxoTransition = do
     trans @(EraRule "UTXOS" era) $
       TRC (UtxosEnv slot pp certState, utxosGovState utxos, stAnnTx)
 
-  case tx ^. isValidTxL of
-    IsValid True ->
+  case tx ^. isPhase2ValidTxL of
+    Phase2Valid ->
       Shelley.updateUTxOState
         pp
         utxos
@@ -583,7 +583,7 @@ utxoTransition = do
         updatedGovState
         (tellEvent . TotalDeposits (hashAnnotated txBody))
         (\a b -> tellEvent (TxUTxODiff a b))
-    IsValid False ->
+    Phase2Invalid ->
       {- utxoKeep = txBody ^. collateralInputsTxBodyL ⋪ utxo -}
       {- utxoDel  = txBody ^. collateralInputsTxBodyL ◁ utxo -}
       let !(utxoKeep, utxoDel) = extractKeys (unUTxO utxo) (txBody ^. collateralInputsTxBodyL)