+++ eras/conway/impl/src/Cardano/Ledger/Conway/Rules/Utxos.hs
@@ -51,10 +51,10 @@ import GHC.Generics (Generic)
 import Lens.Micro ((^.))
 
 data ConwayUtxosPredFailure era
-  = -- | The 'isValid' tag on the transaction is incorrect. The tag given
+  = -- | The 'isPhase2Valid' tag on the transaction is incorrect. The tag given
     --   here is that provided on the transaction (whereas evaluation of the
     --   scripts gives the opposite.). The Text tries to explain why it failed.
-    ValidationTagMismatch IsValid Alonzo.TagMismatchDescription
+    ValidationTagMismatch IsPhase2Valid Alonzo.TagMismatchDescription
   | -- | We could not find all the necessary inputs for a Plutus Script.
     -- Previous PredicateFailure tests should make this impossible, but the
     -- consequences of not detecting this means scripts get dropped, so things
@@ -218,9 +218,9 @@ utxosTransition ::
 utxosTransition =
   judgmentContext >>= \(TRC ((), (), stAnnTx)) -> do
     let tx = stAnnTx ^. txStAnnTxG
-    case tx ^. isValidTxL of
-      IsValid True -> conwayEvalScriptsTxValid
-      IsValid False -> do
+    case tx ^. isPhase2ValidTxL of
+      Phase2Valid -> conwayEvalScriptsTxValid
+      Phase2Invalid -> do
         Babbage.babbageEvalScriptsTxInvalid @era stAnnTx
 
 conwayEvalScriptsTxValid ::