+++ eras/conway/impl/src/Cardano/Ledger/Conway/Translation.hs
@@ -95,11 +95,11 @@ instance TranslateEra ConwayEra (Tx TopTx) where
     txBody <- translateEraThroughCBOR "TxBody" $ tx ^. bodyTxL
     txWits <- translateEraThroughCBOR "TxWitness" $ tx ^. witsTxL
     auxData <- mapM (translateEraThroughCBOR "AuxData") (tx ^. auxDataTxL)
-    let isValidTx = tx ^. isValidTxL
+    let isPhase2ValidTx = tx ^. isPhase2ValidTxL
     pure $
       mkBasicTx txBody
         & witsTxL .~ txWits
-        & isValidTxL .~ isValidTx
+        & isPhase2ValidTxL .~ isPhase2ValidTx
         & auxDataTxL .~ auxData
 
 --------------------------------------------------------------------------------