+++ eras/dijkstra/impl/src/Cardano/Ledger/Dijkstra/Rules/SubUtxo.hs
@@ -66,7 +66,7 @@ data SubUtxoEnv era = SubUtxoEnv
   , suePParams :: PParams era
   , sueCertState :: CertState era
   , sueOriginalUtxo :: UTxO era
-  , sueTopTxIsValid :: IsValid
+  , sueTopTxIsPhase2Valid :: IsPhase2Valid
   }
 
 data DijkstraSubUtxoPredFailure era
@@ -229,7 +229,7 @@ dijkstraSubUtxoTransition ::
   ) =>
   TransitionRule (EraRule "SUBUTXO" era)
 dijkstraSubUtxoTransition = do
-  TRC (SubUtxoEnv slot pp certState originalUtxo (IsValid isValid), utxoState, stAnnTx) <-
+  TRC (SubUtxoEnv slot pp certState originalUtxo topTxIsPhase2Valid, utxoState, stAnnTx) <-
     judgmentContext
   let tx = stAnnTx ^. txStAnnTxG
 
@@ -262,7 +262,7 @@ dijkstraSubUtxoTransition = do
   runTestOnSignal $ validateWrongNetworkInDirectDeposit netId txBody
   runTestOnSignal $ Alonzo.validateWrongNetworkInTxBody netId txBody
 
-  if isValid
+  if topTxIsPhase2Valid == Phase2Valid
     then do
       newState <-
         Shelley.updateUTxOStateNoFees