+++ libs/cardano-ledger-conformance/src/Test/Cardano/Ledger/Conformance/SpecTranslate/Conway/Ledger.hs
@@ -99,7 +99,7 @@ instance SpecTranslate ConwayEra (Tx TopTx ConwayEra) where
       <$> withCtxSpecTransM (txIdTx tx) (toSpecRep (tx ^. bodyTxL))
       <*> toSpecRep (tx ^. witsTxL)
       <*> toSpecRep (tx ^. sizeTxF)
-      <*> toSpecRep (tx ^. isValidTxL)
+      <*> toSpecRep (tx ^. isPhase2ValidTxL)
       <*> toSpecRep (tx ^. auxDataTxL)
 
 instance SpecTranslate ConwayEra (AlonzoStAnnTx TopTx ConwayEra) where